Fixes `mnvgReadPixels()` for Mac.
This commit fixes the issue that `mnvgReadPixels()` doesn't work on Mac. This issue happened because the default storage mode for textures is managed, and needs to synchronize back to the main memory when modified on the GPU. It's resolved by using a blit command encoder before committing the command buffer.